requirements.txt
python-telegram-bot==22.3
websockets==15.0.1
cryptography==45.0.6
mysql-connector-python==9.4.0
numpy
pandas==2.3.1
scipy
matplotlib
statsmodels